Improve Confluence vs Jira detection
Hey guys,

using wappalyzer recently I discovered, that a lot of times websites get marked as Atlassian Confluence AND Jira. However, only the Jira technology is capable of finding a version and the Confluence was not.
This is happening, because both applications have the "ajs-version-number" (which is stated within the Atlassian Jira app rule). As well, the provided html-regex ("Powered by...") is not working - at least on pages I encountered. Tested it as well manually.

So my suggestion for lower false-positives would be:
* Remove "ajs-version-number" from JIRA - because both apps have it.
* Jira pages have one attribute within a meta-tag and the body-tag: "data-version", which we can take advantage of -> therefore changed the html-regex
* Confluence pages have in addition to "Powered by ..." the "Printed by Atlassian Confluence <version>" within the footer section. However, depending on the language the only useable part is "Atlassian Confluence <version>" ("Powered by" gets translated). I guess the "Powered by..." regex could be revised as well, but in my opinion it does make the regex simpler for the same result - this is why I suggest using it.

It is a proposal and I wanted feedback from the community about this. 
If you have any input, let me know and I will work on it. 

Greetings