added removed jquery regex again
The previous "scripts"-regexes (present until v6.6.0) were removed and replaced with "jquery". For me the version detection has declined due to this change. May I ask, what is the reason behind this? If I have missed something, sorry for bothering and feel free to close this proposed change. 

Thanks in advance